requirements.txt
# ultralytics==8.1.33       
# torch>=2.0.0
# torchvision>=0.15.0
# opencv-python
# pillow
# numpy

torch==2.2.0
torchvision==0.17.0
timm==0.9.12
ultralytics==8.0.217
opencv-python==4.9.0.80
numpy==1.24.4
Pillow==10.2.0
scikit-learn==1.3.2
matplotlib==3.8.3  # Phục vụ PIL và một số case liên quan

